changed time type from float to double, for omp_get_wtime() returns a double: stupid me!

+++ include/data_structure.h
@@ -13,7 +13,7 @@ struct _mdsys
 {	
 	int nthreads;
     int natoms, nfi, nsteps;
-    float time_omp;
+    double time_omp;
     double dt, mass, epsilon, sigma, box, rcut;
     double ekin, epot, temp;
     double *rx, *ry, *rz;
